Add VSCode workspace definition
Code has the best extension for writing minecraft datapacks, without
question. But to use it effectively in a repository that's structured
like this one (with a single root and multiple datapacks stored within
it) you need to pull some fuckery, because the extension only recognizes
workspace roots as datapacks if they have a `data` directory. Thankfully
it also supports workspaces, which means you can add each datapack as
another root in the workspace, and ignore them in the root workspace.