Add afk-sleep datapack
afk-sleep allows you to sleep while other players are AFK by adjusting
the playersSleepingPercentage gamerule. The initial state of the gamerule
is used as the percentage of non-afk players that need to sleep in order
for a night to pass, the default being 100%.

For example, with the following conditions...
gamerule playersSleepingPercentage 100
Players online: 3
Players not AFK: 2
... it would require 2 players to pass the night.
If playersSleepingPercentage was set to 50 when the pack was loaded, then
it would require 50% of 2 (so, 1).