Add CI for building with Nix
The intention is to make delevoping on Windows easier - typically the
only way you would be able to on Windows is either using WSL (with very
bad IO speed) or by remoting into another dev machine that's running
macOS or Unix. Instead, this opens a third option: native development
with remote building.