Add system for data pack format extensibility
Fixes #27

This adds misode/mcmeta/summary as a dependency, which includes a file
that tracks the changes to `version.json` over time. It's updated every
time there's an update to the game, including for snapshots. This is now
used as the source of truth for the pack format -> game version mapping.

Because builds will fail if you use an invalid pack format, downstream
users of this flake should also depend on it and set this flake's
dependency to follow it. This allows it to be updated independently, to
avoid fucking up other people's builds with something out of their
control.

The documentation has also been updated to match this, and all examples
of using this as a flake have also been updated. The examples had most
of their comments removed, because with more comments it became harder
to see the note about depending on and overriding mcmeta-summary.

In addition to the above changes, there were also some minor typo fixes.