Remove mcmeta-summary input from examples
Now that CI keeps the input up-to-date, there's no need to manually
override the mcmeta-summary input in your own flake. The only reason
would be to keep the mapping updated while you stay on an older version
of this flake, or if this repo's CI breaks (it's so simple that it
realistically shouldn't ever break).

Fixes #29