Improve testing workflow
There's a new fish script now: build-test.fish
This will copy the built data packs to a Minecraft world datapacks dir.

You must set the environment variable DATAPACK_DIR to a directory. To
make this easier, a file named `.env` is now automatically loaded, and is
ignored by Git.

Use it like so:
    ./build-test.fish

To build and test specific data packs, provide them as arguments.
    ./build-test.fish drop-xp respawn-sickness

The names of the data packs are the Nix build output names.