Improve stability of build-test.fish
By default, it will now copy the source directories to the world instead
of the zipped directories. This is to make debugging easier, as you can
read the contents of the mcfunction files and see what was built, without
having to decompress the zip.

There is now an optional --zip flag that you can pass to copy the zips.