Allow player to edit playersSleepingPercentage directly
Now that the infrastructure for sleeping detection is in place, #60 is
fixable. When no players are sleeping, the playersSleepingPercentage
gamerule is set to -4 (a little easter egg, since my username is clo4).
If this value is not the same as the .playersSleepingPercentage value,
the .relativePercentage value is updated to match that of the gamerule.

This allows players to, in most cases, edit the gamerule directly and
have it "just work" (so long as they don't look at what the value was
and get confused, or attempt to query it).

Fixes #60