Add backwards compatibility for is_sleeping predicate
In 1.19, the "player" field was replaced with the "type_specific" field.
However, the game doesn't check for fields that it doesn't recognize,
which means both the legacy and current fields can be present and the
datapack will still work.

I fully expect this behaviour to change in the future, so I do this
knowing it's going to bite me in the ass. But for now, it gives me strong
compatibility guarantees for new features.