Add new pack.png for afk
An edited image of my keyboard, instead of a bad pixel art illustration
of letters. This is visually a lot better, but it's definitely less
interesting to look at. The other one wasn't good, and this isn't either,
but it feels less amateurish. So until I can finally make one that looks
good, I think I'll stick with this.