Add new afk-kick data pack
This pack isn't tested yet, and I don't intend on testing it yet. This
isn't something that I intend on using myself, but seeing the popularity
of what I'd consider to be fairly simple data packs, I figured I may as
well capitalize on that and start making some low effort ones of my own.