Remove functions to trigger AFK in favor of trigger scoreboard
This is a native way to achieve the same functionality that also allows
admins to grant players the ability to set themselves as AFK if they
want.

This isn't enabled by default for anybody because it gives players the
opportunity for trolling and abuse (e.g. spamming AFK message), but
can be enabled per player (or for the entire server) using a command
such as
    /scoreboard players enable @a afk