Don't send message if player left game while AFK
This fixes an annoying issue where players that left the game while AFK,
such as if you were manning an automated farm with the menu open and
decide to exit the game without "returning" first, would have a message
sent announcing their return from being AFK *immediately* after their
join message. It was redundant, and felt a bit cheap to see that.

The fix is to track whether the player has left the game, and if that
score is anything other than 0 when `back` is called, the player must
have "returned" after leaving the game.