Add option to drop full amount of XP on death
Previously, this data pack always followed the vanilla behaviour, which
was to calculate the number of levels lost, multiply that by 7, and drop
that amount of XP. I personally actually really like this, and want to
keep it as the default. However, it makes sense to support players
recovering the full amount of their XP. Because of the way it's
calculated, it doesn't support dropping a percentage of the player's
levels.