Add no-sprint data pack
This is a simple data pack that allows me to force players to not
sprint. When players cannot sprint, it removes the muscle memory to
start sprinting any time they want to go anywhere.

This is more of an experiment for myself. I don't expect to publish this
any time soon.

There isn't a good way to disable sprinting without any other
consequences, so this data pack uses a very scuffed solution. Applying
blindness will kick players out of sprint, and doing so for only 2 ticks
results in a flicker for 100ms, which is short enough that it doesn't
bother me at all. Human vision mostly corrects for it because it's about
the same duration as blinking.