afk-sit: don't sit if crawling
This commit doesn't interact well with minceraft:scale <1, but that's
unlikely to be an issue.

If the player is crawling, then the space they're in does not have
enough room to sit, so they should be ignored. Unfortunately, the game
does not have a built in way to detect whether a player is crawling, so
the only way I could think of was to check at the player's head and see
if @s could still be detected more than 1 block away. If the entity is
entirely contained in the same block position as the player's eyes, the
player must be crawling.