slow-day: init
This commit introduces the first version of "slow day", a data pack that
increases the length of the in-game day. By default, I plan to have the
day take 20 minutes, and the length of the night will be unchanged.

Unlike existing mods and data packs that do this by either toggling
on/off the doDaylightCycle gamerule or fundamentally changing how long
the day takes, this data pack pauses the time at specific points in the
day.

It is my intention to make this fully user customisable:
- Users should be able to override the durations for the default pauses
- Users should be able to customise when the pauses occur, if they
  really want to.